--- template/__main__.py
@@ -1,5 +1,0 @@
-from cellophane import cellophane
-
-if __name__ == "__main__":
-    _main = cellophane("{label}")
-    _main(prog_name="{prog_name}")
